backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver for FastAPI

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
numpy==1.26.3

# Authentication
bcrypt==4.2.1  # Password hashing
python-jose[cryptography]==3.3.0  # JWT encoding/decoding
python-multipart==0.0.6  # Form data parsing in login

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI